Change price and monetary fields from Integer to Float
- Updated SQLAlchemy models in database_utils/models/crm.py:
  - Product.price: Integer -> Float
  - Order.total: Integer -> Float
  - Invoice.subtotal: Integer -> Float
  - Invoice.tax: Integer -> Float
  - Invoice.total: Integer -> Float

- Updated Pydantic schemas:
  - ProductBase.price: int -> float
  - ProductUpdate.price: Optional[int] -> Optional[float]
  - OrderUpdate.total: Optional[int] -> Optional[float]
  - OrderOut.total: int -> float
  - InvoiceBase (subtotal, tax, total): int -> float
  - InvoiceUpdate (subtotal, tax, total): Optional[int] -> Optional[float]

- Created Alembic migration to alter database columns from INTEGER to FLOAT

This change enables support for decimal prices (e.g., 99.99) instead of
integer-only values, which is essential for accurate financial calculations
and real-world pricing scenarios.

+++ database_utils/schemas/invoice.py
@@ -6,9 +6,9 @@ from uuid import UUID
 
 class InvoiceBase(BaseModel):
     issue_date: datetime
-    subtotal: int
-    tax: int
-    total: int
+    subtotal: float
+    tax: float
+    total: float
     details: dict
 
 
@@ -18,9 +18,9 @@ class InvoiceCreate(InvoiceBase):
 
 class InvoiceUpdate(BaseModel):
     issue_date: Optional[datetime] = None
-    subtotal: Optional[int] = None
-    tax: Optional[int] = None
-    total: Optional[int] = None
+    subtotal: Optional[float] = None
+    tax: Optional[float] = None
+    total: Optional[float] = None
     details: Optional[dict] = None
